requirements.txt
fastapi
uvicorn
gtts
pdfminer.six